requirements.txt
langgraph
langchain
langchain-google-vertexai
google-cloud-bigquery
db-dtypes
pydantic
python-dotenv
pandas
numpy
regex
streamlit
pygments
